Prepare for cinder gate test
We want to be able to run cinderlib functional tests on Cinder gates, so
this patch adds a playbook and a tool to convert from Cinder's config
file to the YAML required by our functional tests.

The cinder-gate-run playbook uses cinderlib_ignore_errors variable to
define whether the playbook should fail on errors or should they just be
ignored.

The playbook stores the test run results on the common logs directory so
they are automatically saved by the Cinder gate jobs.